[Ondro] Monster Facade and Service corrections

+++ service-layer/src/main/java/cz/muni/fi/pa165/hauntedhouses/service/MonsterServiceImpl.java
@@ -4,6 +4,7 @@ import cz.muni.fi.pa165.hauntedhouses.dao.MonsterDao;
 import cz.muni.fi.pa165.hauntedhouses.entity.House;
 import cz.muni.fi.pa165.hauntedhouses.entity.Monster;
 import cz.muni.fi.pa165.hauntedhouses.exceptions.DataManipulationException;
+import cz.muni.fi.pa165.hauntedhouses.exceptions.ServiceExceptionTranslate;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -14,27 +15,41 @@ import javax.inject.Inject;
  * @author Marek Janco
  */
 @Service
+@ServiceExceptionTranslate
 public class MonsterServiceImpl implements MonsterService{
+
     @Inject
     private MonsterDao monsterDao;
 
     @Override
     public void create(Monster m) {
+        if(m == null){
+            throw new IllegalArgumentException("monster is null - cannot create");
+        }
         monsterDao.create(m);
     }
 
     @Override
     public Monster update(Monster m) {
+        if(m == null){
+            throw new IllegalArgumentException("monster is null - cannot update");
+        }
         return monsterDao.update(m);
     }
 
     @Override
     public void delete(Monster m) {
+        if(m == null){
+            throw new IllegalArgumentException("monster is null - cannot delete");
+        }
         monsterDao.delete(m);
     }
 
     @Override
     public Monster findById(Long id) {
+        if(id == null){
+            throw new IllegalArgumentException("id of monster is null");
+        }
         return monsterDao.findById(id);
     }
